Remove base64 and password-based secret providers
This commit eliminates the implementation of specifically base64 and password-based secret providers from the codebase. The changes include the removal of related classes and tests. The remaining "SecretProvider" class is made more generic to handle secrets without specific encoding or encryption methods.

base64 secret provider was about as useful as a chocolate fire guard. I mean, it was purely an unencrypted string.
