+++ src/Aspirate.Commands/Commands/Apply/ApplyOptions.cs
@@ -1,6 +1,6 @@
 namespace Aspirate.Commands.Commands.Apply;
 
-public sealed class ApplyOptions : BaseCommandOptions, IKubernetesOptions, ISecretOptions, IApplyOptions
+public sealed class ApplyOptions : BaseCommandOptions, IKubernetesOptions, IApplyOptions
 {
     public string? InputPath { get; set; }
     public string? KubeContext { get; set; }
