+++ src/Aspirate.Commands/Commands/BaseCommand.cs
@@ -9,15 +9,12 @@ public abstract class BaseCommand<TOptions, TOptionsHandler> : Command
         : base(name, description)
     {
         AddOption(NonInteractiveOption.Instance);
-        AddOption(SecretProviderOption.Instance);
         AddOption(DisableSecretsOption.Instance);
         Handler = CommandHandler.Create<TOptions, IServiceCollection>(ConstructCommand);
     }
 
     private static async Task<int> ConstructCommand(TOptions options, IServiceCollection services)
     {
-        services.RegisterAspirateSecretProvider(options.SecretProvider);
-
         var handler = ActivatorUtilities.CreateInstance<TOptionsHandler>(services.BuildServiceProvider());
 
         var stateOptions = GetStateManagementOptions(options, handler);
