+++ src/Aspirate.Commands/Commands/BaseCommandOptions.cs
@@ -4,7 +4,6 @@ namespace Aspirate.Commands.Commands;
 public abstract class BaseCommandOptions : ICommandOptions
 {
     public bool NonInteractive { get; set; }
-    public SecretProviderType SecretProvider { get; set; }
     public bool DisableSecrets { get; set; }
     public bool DisableState { get; set; }
     public string? SecretPassword { get; set; }
