--- src/Aspirate.Secrets/BaseSecretState.cs
@@ -1,10 +1,0 @@
-namespace Aspirate.Secrets;
-
-public abstract class BaseSecretState
-{
-    [JsonPropertyName("secrets")]
-    public Dictionary<string, Dictionary<string, string>> Secrets { get; set; } = [];
-
-    [JsonPropertyName("secretsVersion")]
-    public int? Version { get; set; } = 0;
-}
