+++ src/Aspirate.Secrets/GlobalUsings.cs
@@ -3,14 +3,12 @@ global using System.Security.Cryptography;
 global using System.Text;
 global using System.Text.Json;
 global using System.Text.Json.Serialization;
-global using Ardalis.SmartEnum;
 global using Aspirate.Secrets.Protectors;
-global using Aspirate.Secrets.Providers;
-global using Aspirate.Secrets.Providers.Base64;
-global using Aspirate.Secrets.Providers.Password;
+global using Aspirate.Shared.Attributes;
 global using Aspirate.Shared.Enums;
 global using Aspirate.Shared.Interfaces.Secrets;
-global using Aspirate.Shared.Literals;
+global using Aspirate.Shared.Interfaces.Services;
+global using Aspirate.Shared.Models.Aspirate;
 global using Aspirate.Shared.Models.AspireManifests;
 global using Aspirate.Shared.Models.AspireManifests.Interfaces;
 global using Microsoft.Extensions.DependencyInjection;
