--- src/Aspirate.Secrets/Providers/Base64/Base64Decrypter.cs
@@ -1,13 +1,0 @@
-namespace Aspirate.Secrets.Providers.Base64;
-
-public class Base64Decrypter : IDecrypter
-{
-    public string DecryptValue(string ciphertext)
-    {
-        var ciphertextBytes = Convert.FromBase64String(ciphertext);
-        return Encoding.UTF8.GetString(ciphertextBytes);
-    }
-
-    public Dictionary<string, string> BulkDecrypt(List<string> ciphertexts) =>
-        ciphertexts.ToDictionary(ciphertext => ciphertext, DecryptValue);
-}
