--- src/Aspirate.Secrets/Providers/Base64/Base64SecretState.cs
@@ -1,3 +1,0 @@
-namespace Aspirate.Secrets.Providers.Base64;
-
-public class Base64SecretState : BaseSecretState, IBase64SecretState;
