+++ src/Aspirate.Secrets/SecretProvider.cs
@@ -1,20 +1,82 @@
 namespace Aspirate.Secrets;
 
-public abstract class BaseSecretProvider<TState>(IFileSystem fileSystem) : ISecretProvider where TState : BaseSecretState, new()
+public class SecretProvider(IFileSystem fileSystem) : ISecretProvider
 {
-    protected readonly JsonSerializerOptions _serializerOptions = new()
+    private const int TagSizeInBytes = 16;
+    private string? _password;
+    private IEncrypter? _encrypter;
+    private IDecrypter? _decrypter;
+    private byte[]? _salt;
+
+    public SecretState? State { get; set; }
+
+    public IEncrypter? Encrypter => _encrypter;
+
+    public IDecrypter? Decrypter => _decrypter;
+
+    public void SetPassword(string password)
+    {
+        _password = password;
+
+        if (_salt is null)
+        {
+            CreateNewSalt();
+        }
+
+        // Derive a key from the passphrase using Pbkdf2 with SHA256, 1 million iterations.
+        using var pbkdf2 = new Rfc2898DeriveBytes(_password, salt: _salt, iterations: 1000000, HashAlgorithmName.SHA256);
+        var key = pbkdf2.GetBytes(32); // AES-256-GCM needs a 32-byte key
+        var crypter = new AesGcmCrypter(key, _salt, TagSizeInBytes);
+
+        _encrypter = crypter;
+        _decrypter = crypter;
+
+        SetPasswordHash();
+    }
+
+    public bool CheckPassword(string password)
+    {
+        using var pbkdf2ToCheck = new Rfc2898DeriveBytes(password, salt: _salt, iterations: 1000000, HashAlgorithmName.SHA256);
+        var passwordToCheckHash = Convert.ToBase64String(pbkdf2ToCheck.GetBytes(32));
+
+        return passwordToCheckHash == State.Hash;
+    }
+
+    public void ProcessAfterStateRestoration()
+    {
+        if (!string.IsNullOrEmpty(_password))
+        {
+            _password = null;
+            _decrypter = null;
+            _encrypter = null;
+        }
+
+        State ??= new();
+
+        _salt = !string.IsNullOrEmpty(State.Salt) ? Convert.FromBase64String(State.Salt) : null;
+    }
+
+    private void CreateNewSalt()
+    {
+        _salt = new byte[12];
+        RandomNumberGenerator.Fill(_salt);
+        State ??= new();
+        State.Salt = Convert.ToBase64String(_salt);
+    }
+
+    private void SetPasswordHash()
+    {
+        using var pbkdf2 = new Rfc2898DeriveBytes(_password, salt: _salt, iterations: 1000000, HashAlgorithmName.SHA256);
+        State.Hash = Convert.ToBase64String(pbkdf2.GetBytes(32));
+    }
+
+
+
+     protected readonly JsonSerializerOptions _serializerOptions = new()
     {
         WriteIndented = true,
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
     };
-
-    public abstract SecretProviderType Type { get; }
-
-    public abstract TState? State { get; protected set; }
-
-    public abstract IEncrypter? Encrypter { get; }
-
-    public abstract IDecrypter? Decrypter { get; }
 
     public void AddSecret(string resourceName, string key, string value)
     {
@@ -41,9 +103,6 @@ public abstract class BaseSecretProvider<TState>(IFileSystem fileSystem) : ISecr
 
     public virtual void TransformStateForStorage() =>
         State.Version++;
-
-    public virtual void ProcessAfterStateRestoration()
-    {}
 
     public void SaveState(string path)
     {
@@ -74,7 +133,7 @@ public abstract class BaseSecretProvider<TState>(IFileSystem fileSystem) : ISecr
         }
 
         var stateJson = fileSystem.File.ReadAllText(path);
-        State = JsonSerializer.Deserialize<TState>(stateJson, _serializerOptions);
+        State = JsonSerializer.Deserialize<SecretState>(stateJson, _serializerOptions);
 
         ProcessAfterStateRestoration();
     }
