+++ src/Aspirate.Secrets/ServiceCollectionExtensions.cs
@@ -1,5 +1,3 @@
-using Aspirate.Shared.Interfaces.Services;
-
 namespace Aspirate.Secrets;
 
 public static class ServiceCollectionExtensions
@@ -10,15 +8,6 @@ public static class ServiceCollectionExtensions
             .AddSingleton<ISecretProtectionStrategy, PostgresPasswordProtector>()
             .AddSingleton<ISecretProtectionStrategy, MsSqlPasswordProtector>();
 
-    public static IServiceCollection RegisterAspirateSecretProvider(this IServiceCollection services, SecretProviderType providerType) =>
-        providerType switch
-        {
-            SecretProviderType.Base64 => services.RegisterSecretProvider<Base64SecretProvider>(),
-            SecretProviderType.Password => services.RegisterSecretProvider<PasswordSecretProvider>(),
-            _ => throw new ArgumentOutOfRangeException(nameof(providerType), providerType, null)
-        };
-
-    private static IServiceCollection RegisterSecretProvider<TImplementation>(this IServiceCollection services)
-        where TImplementation : class, ISecretProvider =>
-            services.AddSingleton<ISecretProvider, TImplementation>();
+    public static IServiceCollection AddAspirateSecretProvider(this IServiceCollection services) =>
+        services.AddSingleton<ISecretProvider, SecretProvider>();
 }
