+++ src/Aspirate.Services/GlobalUsings.cs
@@ -3,8 +3,6 @@ global using System.Security.Cryptography;
 global using System.Text;
 global using System.Text.Json;
 global using System.Text.RegularExpressions;
-global using Aspirate.Secrets.Providers.Base64;
-global using Aspirate.Secrets.Providers.Password;
 global using Aspirate.Services.Implementations;
 global using Aspirate.Shared.Exceptions;
 global using Aspirate.Shared.Extensions;
@@ -28,8 +26,5 @@ global using k8s;
 global using k8s.Models;
 global using Microsoft.Extensions.DependencyInjection;
 global using Spectre.Console;
-global using YamlDotNet.Core;
-global using YamlDotNet.Core.Events;
 global using YamlDotNet.Serialization;
-global using YamlDotNet.Serialization.EventEmitters;
 global using YamlDotNet.Serialization.NamingConventions;
