--- src/Aspirate.Shared/Enums/SecretProviderType.cs
@@ -1,7 +1,0 @@
-namespace Aspirate.Shared.Enums;
-
-public enum SecretProviderType
-{
-    Password = 0,
-    Base64 = 1,
-}
