--- src/Aspirate.Shared/Interfaces/Commands/Contracts/ISecretOptions.cs
@@ -1,6 +1,0 @@
-namespace Aspirate.Shared.Interfaces.Commands.Contracts;
-
-public interface ISecretOptions
-{
-    SecretProviderType SecretProvider { get; set; }
-}
