--- src/Aspirate.Shared/Interfaces/Secrets/IBase64SecretState.cs
@@ -1,8 +1,0 @@
-namespace Aspirate.Shared.Interfaces.Secrets;
-
-public interface IBase64SecretState
-{
-    Dictionary<string, Dictionary<string, string>> Secrets { get; set; }
-
-    int? Version { get; set; }
-}
