+++ tests/Aspirate.Tests/GlobalUsings.cs
@@ -2,14 +2,11 @@ global using System.IO.Abstractions;
 global using System.IO.Abstractions.TestingHelpers;
 global using System.Runtime.CompilerServices;
 global using System.Text.Json;
-global using Aspirate.Cli.Extensions;
 global using Aspirate.Commands.Actions;
 global using Aspirate.Commands.Actions.Containers;
 global using Aspirate.Commands.Actions.Manifests;
 global using Aspirate.Commands.Actions.Secrets;
 global using Aspirate.Processors.Transformation;
-global using Aspirate.Secrets.Providers.Base64;
-global using Aspirate.Secrets.Providers.Password;
 global using Aspirate.Services.Implementations;
 global using Aspirate.Shared.Exceptions;
 global using Aspirate.Shared.Extensions;
